Move the function is_compatible_soc() to a new module
By doing this, the function can be called by platform modules as well as
by the EEPROM module.

This is a preparatory patch for future developments that will expand the
number of platforms on which the bootcount application can operate.
